switch get component query params to use field-input component - move select and boolean types to field-input component - switch get component to use field-input components
